refactor: extract animations module
Moves the DOM-only animation helpers out of js/main.js into the new
js/animations.js module: initScrollReveal, initCardTilt, initCardFlip,
initMagneticButtons, initScroll3D (hero parallax), initSkillBars, and
initTimelineScroll3D.

These functions are pure DOM / Canvas API code — no Three.js
dependency — so the module only imports prefersReducedMotion from
utils.js.

main.js imports them back and re-exports for tests.  207 tests still
pass.
